requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
pandas
openpyxl
waitress
pyodbc
sqlalchemy